Import common kernel 5.15.148 from S918NKSU5CXK4, SM-S918N_KOR_14_Opensource.zip

+++ drivers/block/zram/zram_drv.h
@@ -196,7 +196,7 @@ struct zram {
 #ifdef CONFIG_ZRAM_LRU_WRITEBACK
 	struct task_struct *wbd;
 	wait_queue_head_t wbd_wait;
-	u8 *wb_table;
+	u16 *wb_table;
 	unsigned long *chunk_bitmap;
 	unsigned long nr_lru_pages;
 	bool wbd_running;
